Introduce class CodeGenTypes
