+++ polygeist/tools/cgeist/Lib/CGCall.cc
@@ -8,9 +8,11 @@
 
 #include "TypeUtils.h"
 #include "clang-mlir.h"
+#include "utils.h"
+
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "mlir/Dialect/SYCL/IR/SYCLOps.h"
-#include "utils.h"
+#include "mlir/Target/LLVMIR/TypeFromLLVM.h"
 
 #define DEBUG_TYPE "CGCall"
 
@@ -21,6 +23,10 @@ using namespace mlir::func;
 using namespace mlirclang;
 
 extern llvm::cl::opt<bool> GenerateAllSYCLFuncs;
+
+/******************************************************************************/
+/*                           Utility Functions                                */
+/******************************************************************************/
 
 /// Try to typecast the caller arg of type MemRef to fit the corresponding
 /// callee arg type. We only deal with the cast where src and dst have the same
@@ -88,6 +94,10 @@ static void castCallerArgs(mlir::func::FuncOp callee,
   }
 }
 
+/******************************************************************************/
+/*                               MLIRScanner                                  */
+/******************************************************************************/
+
 ValueCategory MLIRScanner::CallHelper(
     mlir::func::FuncOp tocall, QualType objType,
     ArrayRef<std::pair<ValueCategory, clang::Expr *>> arguments,
@@ -102,7 +112,7 @@ ValueCategory MLIRScanner::CallHelper(
     GD = GlobalDecl(CC, CXXDtorType::Dtor_Complete);
   else
     GD = GlobalDecl(callee);
-  const CodeGen::CGFunctionInfo &FI =
+  const clang::CodeGen::CGFunctionInfo &FI =
       Glob.getCGM().getTypes().arrangeGlobalDeclaration(GD);
   auto FuncInfos = FI.arguments();
 
@@ -144,7 +154,7 @@ ValueCategory MLIRScanner::CallHelper(
 
     bool isArray = false;
     QualType aType = (i == 0 && a == nullptr) ? objType : a->getType();
-    auto expectedType = Glob.getMLIRType(aType, &isArray);
+    mlir::Type expectedType = Glob.getTypes().getMLIRType(aType, &isArray);
 
     LLVM_DEBUG({
       llvm::dbgs() << "aType: " << aType << "\n";
@@ -173,7 +183,8 @@ ValueCategory MLIRScanner::CallHelper(
         assert(arg.isReference);
 
         auto mt =
-            Glob.getMLIRType(
+            Glob.getTypes()
+                .getMLIRType(
                     Glob.getCGM().getContext().getLValueReferenceType(aType))
                 .cast<MemRefType>();
 
@@ -208,9 +219,10 @@ ValueCategory MLIRScanner::CallHelper(
                                   mt.getMemorySpace()),
             alloc);
       } else {
-        if (FuncInfos[i].info.getKind() == CodeGen::ABIArgInfo::Indirect ||
+        if (FuncInfos[i].info.getKind() ==
+                clang::CodeGen::ABIArgInfo::Indirect ||
             FuncInfos[i].info.getKind() ==
-                CodeGen::ABIArgInfo::IndirectAliased) {
+                clang::CodeGen::ABIArgInfo::IndirectAliased) {
           OpBuilder abuilder(builder.getContext());
           abuilder.setInsertionPointToStart(allocationScope);
           auto Ty = arg.getValue(builder).getType();
@@ -251,7 +263,7 @@ ValueCategory MLIRScanner::CallHelper(
     } else {
       assert(arg.isReference);
 
-      expectedType = Glob.getMLIRType(
+      expectedType = Glob.getTypes().getMLIRType(
           Glob.getCGM().getContext().getLValueReferenceType(aType));
 
       val = arg.val;
@@ -290,13 +302,13 @@ ValueCategory MLIRScanner::CallHelper(
 
   bool isArrayReturn = false;
   if (!retReference)
-    Glob.getMLIRType(retType, &isArrayReturn);
+    Glob.getTypes().getMLIRType(retType, &isArrayReturn);
 
   mlir::Value alloc;
   if (isArrayReturn) {
-    llvm::dbgs() << "at line " << __LINE__ << "\n";
     auto mt =
-        Glob.getMLIRType(
+        Glob.getTypes()
+            .getMLIRType(
                 Glob.getCGM().getContext().getLValueReferenceType(retType))
             .cast<MemRefType>();
 
@@ -487,7 +499,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
                             : rhs->getExprOperand()->getType();
           llvm::Constant *LC = Glob.getCGM().GetAddrOfRTTIDescriptor(LT);
           llvm::Constant *RC = Glob.getCGM().GetAddrOfRTTIDescriptor(RT);
-          auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+          auto postTy = Glob.getTypes()
+                            .getMLIRType(expr->getType())
+                            .cast<mlir::IntegerType>();
           return ValueCategory(
               builder.create<arith::ConstantIntOp>(loc, LC == RC, postTy),
               false);
@@ -566,6 +580,8 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
       }
     }
 
+  mlir::LLVM::TypeFromLLVMIRTranslator typeTranslator(*module->getContext());
+
   auto getLLVM = [&](Expr *E) -> mlir::Value {
     auto sub = Visit(E);
     if (!sub.val) {
@@ -589,13 +605,14 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
     }
 
     bool isArray = false;
-    Glob.getMLIRType(E->getType(), &isArray);
+    Glob.getTypes().getMLIRType(E->getType(), &isArray);
 
     if (isArray) {
       assert(sub.isReference);
       auto mt =
-          Glob.getMLIRType(Glob.getCGM().getContext().getLValueReferenceType(
-                               E->getType()))
+          Glob.getTypes()
+              .getMLIRType(Glob.getCGM().getContext().getLValueReferenceType(
+                  E->getType()))
               .cast<MemRefType>();
       auto shape = std::vector<int64_t>(mt.getShape());
       assert(shape.size() == 2);
@@ -605,9 +622,10 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
       auto one = abuilder.create<ConstantIntOp>(loc, 1, 64);
       auto alloc = abuilder.create<mlir::LLVM::AllocaOp>(
           loc,
-          LLVM::LLVMPointerType::get(Glob.getTypeTranslator().translateType(
-                                         anonymize(getLLVMType(E->getType()))),
-                                     0),
+          LLVM::LLVMPointerType::get(
+              typeTranslator.translateType(
+                  anonymize(getLLVMType(E->getType(), Glob.getCGM()))),
+              0),
           one, 0);
       ValueCategory(alloc, /*isRef*/ true)
           .store(builder, sub, /*isArray*/ isArray);
@@ -615,8 +633,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
     }
     auto val = sub.getValue(builder);
     if (auto mt = val.getType().dyn_cast<MemRefType>()) {
-      auto nt = Glob.getTypeTranslator()
-                    .translateType(anonymize(getLLVMType(E->getType())))
+      auto nt = typeTranslator
+                    .translateType(
+                        anonymize(getLLVMType(E->getType(), Glob.getCGM())))
                     .cast<LLVM::LLVMPointerType>();
       assert(nt.getAddressSpace() == mt.getMemorySpaceAsInt() &&
              "val does not have the same memory space as nt");
@@ -637,7 +656,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
            sr->getDecl()->getName() == "__nv_powi" ||
            sr->getDecl()->getName() == "__nv_powi" ||
            sr->getDecl()->getName() == "powf")) {
-        auto mlirType = getMLIRType(expr->getType());
+        mlir::Type mlirType = Glob.getTypes().getMLIRType(expr->getType());
         std::vector<mlir::Value> args;
         for (auto *a : expr->arguments()) {
           args.push_back(Visit(a).getValue(builder));
@@ -660,7 +679,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         auto V = Visit(expr->getArg(0));
         assert(V.isReference);
         mlir::Value val = V.val;
-        auto T = getMLIRType(expr->getType());
+        mlir::Type T = Glob.getTypes().getMLIRType(expr->getType());
         if (T == val.getType())
           return ValueCategory(val, /*isRef*/ false);
         if (T.isa<LLVM::LLVMPointerType>()) {
@@ -745,7 +764,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         if (Glob.getFunctions().find(name) == Glob.getFunctions().end()) {
           std::vector<mlir::Type> types{V0.getType(), V1.getType()};
 
-          auto RT = getMLIRType(expr->getType());
+          mlir::Type RT = Glob.getTypes().getMLIRType(expr->getType());
           std::vector<mlir::Type> rettypes{RT};
           mlir::OpBuilder mbuilder(module->getContext());
           auto funcType = mbuilder.getFunctionType(types, rettypes);
@@ -773,7 +792,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         if (Glob.getFunctions().find(name) == Glob.getFunctions().end()) {
           std::vector<mlir::Type> types{V0.getType()};
 
-          auto RT = getMLIRType(expr->getType());
+          mlir::Type RT = Glob.getTypes().getMLIRType(expr->getType());
           std::vector<mlir::Type> rettypes{RT};
           mlir::OpBuilder mbuilder(module->getContext());
           auto funcType = mbuilder.getFunctionType(types, rettypes);
@@ -808,7 +827,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
                         ? CmpFPredicate::OEQ
                         : CmpFPredicate::ONE;
         mlir::Value FCmp = builder.create<CmpFOp>(loc, Pred, Fabs, Infinity);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, FCmp);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -817,7 +838,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
            sr->getDecl()->getName() == "__nv_isnanf")) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value Eq = builder.create<CmpFOp>(loc, CmpFPredicate::UNO, V, V);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, Eq);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -839,7 +862,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
             builder.create<CmpFOp>(loc, CmpFPredicate::UGE, Abs, SmallestV);
         V = builder.create<AndIOp>(loc, Eq, IsLessThanInf);
         V = builder.create<AndIOp>(loc, V, IsNormal);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -851,7 +876,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value BC = builder.create<BitcastOp>(loc, ITy, V);
         auto ZeroV = builder.create<ConstantIntOp>(loc, 0, ITy);
         V = builder.create<CmpIOp>(loc, CmpIPredicate::slt, BC, ZeroV);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -860,7 +887,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::OGT, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -869,7 +898,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::OGE, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -878,7 +909,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::OLT, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -887,7 +920,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::OLE, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -896,7 +931,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::ONE, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -905,7 +942,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V = getLLVM(expr->getArg(0));
         mlir::Value V2 = getLLVM(expr->getArg(1));
         V = builder.create<CmpFOp>(loc, CmpFPredicate::UNO, V, V2);
-        auto postTy = getMLIRType(expr->getType()).cast<mlir::IntegerType>();
+        auto postTy = Glob.getTypes()
+                          .getMLIRType(expr->getType())
+                          .cast<mlir::IntegerType>();
         mlir::Value res = builder.create<ExtUIOp>(loc, postTy, V);
         return ValueCategory(res, /*isRef*/ false);
       }
@@ -941,7 +980,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
         mlir::Value V2 = getLLVM(expr->getArg(1));
         auto name = sr->getDecl()->getName().substr(5).str();
         std::vector<mlir::Type> types{V.getType(), V2.getType()};
-        auto RT = getMLIRType(expr->getType());
+        mlir::Type RT = Glob.getTypes().getMLIRType(expr->getType());
 
         std::vector<mlir::Type> rettypes{RT};
 
@@ -1211,9 +1250,9 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
               }
 
               bool dstArray = false;
-              Glob.getMLIRType(QualType(elem, 0), &dstArray);
+              Glob.getTypes().getMLIRType(QualType(elem, 0), &dstArray);
               bool srcArray = false;
-              Glob.getMLIRType(QualType(selem, 0), &srcArray);
+              Glob.getTypes().getMLIRType(QualType(selem, 0), &srcArray);
               auto elemSize = getTypeSize(QualType(elem, 0));
               if (srcArray && !dstArray)
                 elemSize = getTypeSize(QualType(selem, 0));
@@ -1259,7 +1298,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
 
               if (dstArray) {
                 std::vector<mlir::Value> start = {getConstantIndex(0)};
-                auto mt = Glob.getMLIRType(Glob.getCGM().getContext().getPointerType(
+                auto mt = Glob.getTypes().getMLIRType(Glob.getCGM().getContext().getPointerType(
                                                QualType(elem, 0)))
                               .cast<MemRefType>();
                 auto shape = std::vector<int64_t>(mt.getShape());
@@ -1272,7 +1311,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
                     &affineOp.getLoopBody().front());
                 if (srcArray) {
                   auto smt =
-                      Glob.getMLIRType(Glob.getCGM().getContext().getPointerType(
+                      Glob.getTypes().getMLIRType(Glob.getCGM().getContext().getPointerType(
                                            QualType(elem, 0)))
                           .cast<MemRefType>();
                   auto sshape = std::vector<int64_t>(smt.getShape());
@@ -1289,7 +1328,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
               } else {
                 if (srcArray) {
                   auto smt =
-                      Glob.getMLIRType(Glob.getCGM().getContext().getPointerType(
+                      Glob.getTypes().getMLIRType(Glob.getCGM().getContext().getPointerType(
                                            QualType(selem, 0)))
                           .cast<MemRefType>();
                   auto sshape = std::vector<int64_t>(smt.getShape());
@@ -1341,7 +1380,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
               // at the end of the loop.
               builder.setInsertionPoint(oldblock, oldpoint);
 
-              auto retTy = getMLIRType(expr->getType());
+              mlir::Type retTy = Glob.getTypes().getMLIRType(expr->getType());
               if (sr->getDecl()->getName() == "__builtin_memcpy" ||
                   retTy.isa<LLVM::LLVMPointerType>()) {
                 if (dst.getType().isa<MemRefType>()) {
@@ -1410,7 +1449,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
             if (dst.getType().isa<MemRefType>()) {
 
               bool dstArray = false;
-              auto melem = Glob.getMLIRType(elem, &dstArray);
+              mlir::Type melem = Glob.getTypes().getMLIRType(elem, &dstArray);
               mlir::Value toStore;
               if (melem.isa<mlir::IntegerType>())
                 toStore = builder.create<ConstantIntOp>(loc, 0, melem);
@@ -1440,7 +1479,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
               if (dstArray) {
                 std::vector<mlir::Value> start = {getConstantIndex(0)};
                 auto mt =
-                    Glob.getMLIRType(Glob.getCGM().getContext().getPointerType(elem))
+                    Glob.getTypes().getMLIRType(Glob.getCGM().getContext().getPointerType(elem))
                         .cast<MemRefType>();
                 auto shape = std::vector<int64_t>(mt.getShape());
                 auto affineOp = builder.create<scf::ForOp>(
@@ -1457,7 +1496,7 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
               // at the end of the loop.
               builder.setInsertionPoint(oldblock, oldpoint);
 
-              auto retTy = getMLIRType(expr->getType());
+              mlir::Type retTy = Glob.getTypes().getMLIRType(expr->getType());
               return ValueCategory(builder.create<ConstantIntOp>(loc, 0, retTy),
                                    /*isReference*/ false);
             }
@@ -1546,8 +1585,8 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
                        .getResult(0);
         } else {
           args.insert(args.begin(), getLLVM(expr->getCallee()));
-          SmallVector<mlir::Type> RTs = {Glob.getTypeTranslator().translateType(
-              anonymize(getLLVMType(expr->getType())))};
+          SmallVector<mlir::Type> RTs = {typeTranslator.translateType(
+              anonymize(getLLVMType(expr->getType(), Glob.getCGM())))};
           if (RTs[0].isa<LLVM::LLVMVoidType>())
             RTs.clear();
           called =
@@ -1574,8 +1613,8 @@ ValueCategory MLIRScanner::VisitCallExpr(clang::CallExpr *expr) {
       auto CT = expr->getType();
       if (isReference)
         CT = Glob.getCGM().getContext().getLValueReferenceType(CT);
-      SmallVector<mlir::Type> RTs = {
-          Glob.getTypeTranslator().translateType(anonymize(getLLVMType(CT)))};
+      SmallVector<mlir::Type> RTs = {typeTranslator.translateType(
+          anonymize(getLLVMType(CT, Glob.getCGM())))};
 
       auto ft = args[0]
                     .getType()
