+++ polygeist/tools/cgeist/Lib/TypeUtils.cc
@@ -7,6 +7,16 @@
 //===----------------------------------------------------------------------===//
 
 #include "TypeUtils.h"
+#include "CodeGenTypes.h"
+
+#include "clang/../../lib/CodeGen/CodeGenModule.h"
+#include "clang/AST/Decl.h"
+#include "clang/AST/DeclTemplate.h"
+#include "clang/AST/Type.h"
+
+#include "mlir/Dialect/SYCL/IR/SYCLOps.h"
+#include "mlir/IR/Types.h"
+
 #include "llvm/IR/DerivedTypes.h"
 #include "llvm/Support/Casting.h"
 
@@ -59,6 +69,101 @@ Type *anonymize(Type *T) {
     return StructType::get(ST->getContext(), V, ST->isPacked());
   }
   return T;
+}
+
+mlir::Type getSYCLType(const clang::RecordType *RT,
+                       mlirclang::CodeGen::CodeGenTypes &CGT) {
+  const auto *RD = RT->getAsRecordDecl();
+  llvm::SmallVector<mlir::Type, 4> Body;
+
+  for (const auto *Field : RD->fields())
+    Body.push_back(CGT.getMLIRType(Field->getType()));
+
+  if (const auto *CTS =
+          llvm::dyn_cast<clang::ClassTemplateSpecializationDecl>(RD)) {
+    if (CTS->getName() == "range") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::RangeType::get(CGT.getModule()->getContext(), Dim);
+    }
+    if (CTS->getName() == "nd_range") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::NdRangeType::get(CGT.getModule()->getContext(), Dim,
+                                          Body);
+    }
+    if (CTS->getName() == "array") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::ArrayType::get(CGT.getModule()->getContext(), Dim,
+                                        Body);
+    }
+    if (CTS->getName() == "id") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::IDType::get(CGT.getModule()->getContext(), Dim);
+    }
+    if (CTS->getName() == "accessor_common") {
+      return mlir::sycl::AccessorCommonType::get(CGT.getModule()->getContext());
+    }
+    if (CTS->getName() == "accessor") {
+      const auto Type =
+          CGT.getMLIRType(CTS->getTemplateArgs().get(0).getAsType());
+      const auto Dim =
+          CTS->getTemplateArgs().get(1).getAsIntegral().getExtValue();
+      const auto MemAccessMode = static_cast<mlir::sycl::MemoryAccessMode>(
+          CTS->getTemplateArgs().get(2).getAsIntegral().getExtValue());
+      const auto MemTargetMode = static_cast<mlir::sycl::MemoryTargetMode>(
+          CTS->getTemplateArgs().get(3).getAsIntegral().getExtValue());
+      return mlir::sycl::AccessorType::get(CGT.getModule()->getContext(), Type,
+                                           Dim, MemAccessMode, MemTargetMode,
+                                           Body);
+    }
+    if (CTS->getName() == "AccessorImplDevice") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::AccessorImplDeviceType::get(
+          CGT.getModule()->getContext(), Dim, Body);
+    }
+    if (CTS->getName() == "item") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      const auto Offset =
+          CTS->getTemplateArgs().get(1).getAsIntegral().getExtValue();
+      return mlir::sycl::ItemType::get(CGT.getModule()->getContext(), Dim,
+                                       Offset, Body);
+    }
+    if (CTS->getName() == "ItemBase") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      const auto Offset =
+          CTS->getTemplateArgs().get(1).getAsIntegral().getExtValue();
+      return mlir::sycl::ItemBaseType::get(CGT.getModule()->getContext(), Dim,
+                                           Offset, Body);
+    }
+    if (CTS->getName() == "nd_item") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::NdItemType::get(CGT.getModule()->getContext(), Dim,
+                                         Body);
+    }
+    if (CTS->getName() == "group") {
+      const auto Dim =
+          CTS->getTemplateArgs().get(0).getAsIntegral().getExtValue();
+      return mlir::sycl::GroupType::get(CGT.getModule()->getContext(), Dim,
+                                        Body);
+    }
+  }
+
+  llvm_unreachable("SYCL type not handle (yet)");
+}
+
+llvm::Type *getLLVMType(const clang::QualType QT,
+                        clang::CodeGen::CodeGenModule &CGM) {
+  if (QT->isVoidType())
+    return llvm::Type::getVoidTy(CGM.getModule().getContext());
+
+  return CGM.getTypes().ConvertType(QT);
 }
 
 } // namespace mlirclang
